fix: 修复 CI lint 与续传边界
- errcheck 全绿：init 中陈旧 meta 删除失败现作为错误返回（防污染
  续传）；新增 meta 存在但 partial 缺失的守卫，使完成路径上残留
  meta 的失败清理不再可能在下一次 Start 时拼接出损坏文件；其余
  尽力清理路径加 nolint 注释并说明兜底机制
- misspell：两处 cancelled → canceled
- unconvert/unused：去除多余类型转换与未用测试 helper
- 清除 golangci-lint 全部告警（本机 v1.62.2 复检 0 issues）

+++ limit_test.go
@@ -72,7 +72,7 @@ func TestThrottleWriterWaits(t *testing.T) {
 	}
 	// 再写 128 KiB：按速率需约 1s。
 	start := time.Now()
-	if _, err := tw.WriteAt(make([]byte, rate), int64(rate)); err != nil {
+	if _, err := tw.WriteAt(make([]byte, rate), rate); err != nil {
 		t.Fatalf("WriteAt: %v", err)
 	}
 	elapsed := time.Since(start)
